+++ src/extensions/default/Microsoft.DotNet.UpgradeAssistant.Steps.Packages/DependencyAnalyzerRunner.cs
@@ -14,15 +14,15 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Packages
     {
         private readonly IPackageRestorer _packageRestorer;
         private readonly IEnumerable<IDependencyAnalyzer> _packageAnalyzers;
-
         private readonly ILogger<DependencyAnalyzerRunner> _logger;
 
-        public DependencyAnalyzerRunner(IPackageRestorer packageRestorer,
+        public DependencyAnalyzerRunner(
+            IPackageRestorer packageRestorer,
             IEnumerable<IDependencyAnalyzer> packageAnalyzers,
             ILogger<DependencyAnalyzerRunner> logger)
         {
             _packageRestorer = packageRestorer ?? throw new ArgumentNullException(nameof(packageRestorer));
-            _packageAnalyzers = packageAnalyzers ?? throw new ArgumentNullException(nameof(packageAnalyzers));
+            _packageAnalyzers = packageAnalyzers?.OrderyByPrecedence() ?? throw new ArgumentNullException(nameof(packageAnalyzers));
             _logger = logger;
         }
 
@@ -41,6 +41,7 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Packages
             foreach (var analyzer in _packageAnalyzers)
             {
                 _logger.LogDebug("Analyzing packages with {AnalyzerName}", analyzer.Name);
+
                 try
                 {
                     await analyzer.AnalyzeAsync(projectRoot, analysisState, token).ConfigureAwait(false);